# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Plataforma de Cursos",
        "description": "API para gestionar cursos, instructores, lecciones, comentarios, calificaciones y favoritos.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/instructors": {
            "get": {
                "tags": [
                    "Instructors"
                ],
                "summary": "Lista instructores",
                "operationId": "listInstructors",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado de instructores"
                    }
                }
            },
            "post": {
                "tags": [
                    "Instructors"
                ],
                "summary": "Crea un instructor",
                "operationId": "storeInstructor",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Ana Martinez"
                                    },
                                    "headline": {
                                        "type": "string",
                                        "example": "Senior Backend Engineer"
                                    },
                                    "bio": {
                                        "type": "string",
                                        "example": "Instructora con experiencia en Laravel y arquitectura de software."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Instructor creado"
                    },
                    "422": {
                        "description": "Error de validacion"
                    }
                }
            }
        },
        "/api/instructors/{instructor}": {
            "get": {
                "tags": [
                    "Instructors"
                ],
                "summary": "Obtiene un instructor",
                "operationId": "showInstructor",
                "parameters": [
                    {
                        "name": "instructor",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Instructor encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Instructors"
                ],
                "summary": "Actualiza un instructor",
                "operationId": "updateInstructor",
                "parameters": [
                    {
                        "name": "instructor",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Ana Martinez Updated"
                                    },
                                    "headline": {
                                        "type": "string",
                                        "example": "Lead Backend Engineer"
                                    },
                                    "bio": {
                                        "type": "string",
                                        "example": "Actualizacion del perfil del instructor."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Instructor actualizado"
                    },
                    "422": {
                        "description": "Error de validacion"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Instructors"
                ],
                "summary": "Elimina un instructor",
                "operationId": "destroyInstructor",
                "parameters": [
                    {
                        "name": "instructor",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Instructor eliminado"
                    }
                }
            }
        },
...
    },
...
}